+++ lib/lexer/values/FloatValue.cpp
@@ -2,7 +2,7 @@
 
 #include <sstream>
 
-FloatValue::FloatValue(long double x) : v(x) {
+FloatValue::FloatValue(double x) : v(x) {
 }
 
 std::unique_ptr<Value> FloatValue::Clone() const {
@@ -11,7 +11,7 @@ std::unique_ptr<Value> FloatValue::Clone() const {
 
 std::string FloatValue::ToString() const {
   std::ostringstream os;
-  os << static_cast<double>(v);
+  os << v;
   return os.str();
 }
